requirements.txt
aioredis==1.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.12
fastapi-cache2==0.2.2
fonttools==4.56.0
h11==0.14.0
hiredis==3.1.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
kiwisolver==1.4.8
limits==4.4.1
matplotlib==3.10.1
numpy==2.2.4
openai==1.68.2
packaging==24.2
pendulum==3.0.0
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
requests==2.32.3
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.1
time-machine==2.16.0
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
